requirements.txt
# Core dependencies
pdfplumber==0.10.3
openai==1.59.7
python-dotenv==1.0.0

# API dependencies
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
requests==2.32.3

# PDF/DOCX export
reportlab==4.2.5
python-docx==1.1.2
markdown2==2.5.1

# Vector database
chromadb==0.5.23
numpy==1.26.4
